r legger saksbehandlingstider, sto og chat i egen komponent
